Split test_bench_serving_1gpu.py into 2 parts and add 2 partitions
- Split test_bench_serving_1gpu.py into:
  - test_bench_serving_1gpu_part1.py (LLM + LoRA tests, est_time=1000s)
  - test_bench_serving_1gpu_part2.py (VLM + Score + Embeddings, est_time=900s)
- Add 2 partitions to stage-b-test-large-1-gpu-performance job
- This fixes timeout issue where single file took 31min but had 30min timeout
